Extract common parts from download-runtime script
