+++ src/bindings/js/node/package.json
@@ -23,7 +23,7 @@
     "test:e2e": "mocha ./tests/e2e/electron-app.test.js",
     "tsc": "tsc",
     "postinstall": "npm run install_runtime",
-    "download_runtime": "node ./scripts/download_runtime.js",
+    "download_runtime": "node ./scripts/download-runtime.js",
     "install_runtime": "npm run download_runtime -- --ignore-if-exists"
   },
   "devDependencies": {
