+++ src/bindings/js/node/scripts/lib/binary-manager.js
@@ -1,18 +1,15 @@
 const os = require('node:os');
 const tar = require('tar-fs');
 const path = require('node:path');
-const https = require('node:https');
 const gunzip = require('gunzip-maybe');
 const fs = require('node:fs/promises');
-const { HttpsProxyAgent } = require('https-proxy-agent');
-const { createReadStream, createWriteStream } = require('node:fs');
+const { createReadStream } = require('node:fs');
 
-const packageJson = require('../package.json');
-
-const codeENOENT = 'ENOENT';
+const { downloadFile, checkIfPathExists, removeDirectory } = require('./utils');
 
 class BinaryManager {
-  constructor(version, binaryConfig) {
+  constructor(packageRoot, version, binaryConfig) {
+    this.packageRoot = packageRoot;
     this.version = version;
     this.binaryConfig = binaryConfig;
   }
@@ -48,11 +45,25 @@ class BinaryManager {
   getDestinationPath() {
     const modulePath = this.binaryConfig['module_path'];
 
-    return path.resolve(__dirname, '..', modulePath);
+    return path.resolve(this.packageRoot, modulePath);
   }
 
-  static async prepareBinary(version, binaryConfig, options) {
-    const binaryManager = new this(version, binaryConfig);
+  /**
+   * Prepares the binary by downloading and extracting the OpenVINO runtime archive.
+   *
+   * @param {string} packageRoot - The root directory of the package.
+   * @param {string} version - The version of the binary.
+   * @param {Object} binaryConfig - The configuration object for the binary.
+   * @param {Object} options - The options for preparing the binary.
+   * @param {boolean} options.force - Whether to force the download if the directory already exists.
+   * @param {boolean} options.ignoreIfExists - Whether to ignore the download if the directory already exists.
+   * @param {string} [options.proxy] - The proxy to use for downloading the file.
+   * @throws {Error} If the directory already exists and the force option is not set.
+   * @throws {Error} If the download or extraction fails.
+   * @returns {Promise<void>} A promise that resolves when the binary is prepared.
+   */
+  static async prepareBinary(packageRoot, version, binaryConfig, options) {
+    const binaryManager = new this(packageRoot, version, binaryConfig);
     const destinationPath = binaryManager.getDestinationPath();
     const isRuntimeDirectoryExists = await checkIfPathExists(destinationPath);
 
@@ -101,6 +112,16 @@ class BinaryManager {
     }
   }
 
+  /**
+   * Checks if the current platform and architecture are compatible.
+   *
+   * Supported platforms: 'win32', 'linux', 'darwin'.
+   * Supported architectures: 'arm64', 'armhf', 'x64'.
+   *
+   * If the platform or architecture is not supported, an error message is logged to the console.
+   *
+   * @returns {boolean} Returns true if the platform and architecture are compatible, otherwise false.
+   */
   static isCompatible() {
     const missleadings = [];
     const platform = os.platform();
@@ -147,123 +168,4 @@ class BinaryManager {
   }
 }
 
-if (require.main === module) main();
-
 module.exports = BinaryManager;
-
-async function main() {
-  if (!BinaryManager.isCompatible()) process.exit(1);
-
-  const force = process.argv.includes('-f');
-  const ignoreIfExists = process.argv.includes('--ignore-if-exists');
-
-  const { env } = process;
-  const proxy = env.http_proxy || env.HTTP_PROXY || env.npm_config_proxy;
-
-  await BinaryManager.prepareBinary(
-    packageJson.version,
-    packageJson.binary,
-    { force, ignoreIfExists, proxy },
-  );
-}
-
-/**
- * Remove directory and its content.
- *
- * @async
- * @function removeDirectory
- * @param {string} path - The directory path.
- * @returns {Promise<void>}
- */
-async function removeDirectory(path) {
-  try {
-    console.log(`Removing ${path}`);
-    await fs.rm(path, { recursive: true, force: true });
-  } catch (error) {
-    if (error.code === codeENOENT) console.log(`Path: ${path} doesn't exist`);
-
-    throw error;
-  }
-}
-
-/**
- * Check if path exists.
- *
- * @async
- * @function checkIfPathExists
- * @param {string} path - The path to directory or file.
- * @returns {Promise<boolean>}
- */
-async function checkIfPathExists(path) {
-  try {
-    await fs.access(path);
-    return true;
-  } catch (error) {
-    if (error.code === codeENOENT) {
-      return false;
-    }
-    throw error;
-  }
-}
-
-/**
- * Download file by URL and save it to the destination path.
- *
- * @function downloadFile
- * @param {string} url - The file URL.
- * @param {string} filename - The filename of result file.
- * @param {string} destination - The destination path of result file.
- * @param {string} [proxy=null] - (Optional) The proxy URL.
- * @returns {Promise<string>} - Path to downloaded file.
- */
-function downloadFile(url, destination, filename, proxy = null) {
-  console.log(`Downloading file by link: ${url} to ${destination}`
-    + `with filename: ${filename}`);
-
-  const timeout = 5000;
-  const fullPath = path.resolve(destination, filename);
-  const file = createWriteStream(fullPath);
-
-  if (new URL(url).protocol === 'http')
-    throw new Error('Http link doesn\'t support');
-
-  let agent;
-
-  if (proxy) {
-    agent = new HttpsProxyAgent(proxy);
-    console.log(`Proxy agent is configured with '${proxy}'.`);
-  }
-
-  return new Promise((resolve, reject) => {
-    file.on('error', (error) => {
-      reject(`Failed to open file stream: ${error}.`);
-    });
-
-    console.log(`Download file by link: ${url}`);
-
-    const request = https.get(url, { agent }, (res) => {
-      const { statusCode } = res;
-
-      if (statusCode !== 200) {
-        return reject(`Server returned status code ${statusCode}.`);
-      }
-
-      res.pipe(file);
-
-      file.on('finish', () => {
-        file.close();
-        console.log(`File was successfully downloaded to '${fullPath}'.`);
-        resolve(fullPath);
-      });
-    });
-
-    request.on('error', (error) => {
-      reject(`Failed to send request: ${error}.`);
-    });
-
-    request.setTimeout(timeout, () => {
-      request.destroy();
-      reject(`Request was timed out after ${timeout} ms.`);
-    });
-  });
-}
